+++ typing-tests/tests/error-missing-response-header.ts
@@ -1,11 +1,10 @@
-import { Response, RouteHandler, routeHandler } from 'typera-koa'
+import { Response, Route, route } from 'typera-koa'
 
-export const handler: RouteHandler<Response.Ok<
-  string,
-  { 'Content-Type': 'application/json' }
->> = routeHandler()(_req => {
+export const handler: Route<
+  Response.Ok<string, { 'Content-Type': 'application/json' }>
+> = route.get('/').handler(_req => {
   return Response.ok('foo')
 })
 
 // Expected error:
-// Type 'RouteHandler<RequestBase, Response<200, string, undefined>>' is not assignable to type 'RouteHandler<RequestBase, Response<200, string, { 'Content-Type': "application/json"; }>>'.
+// Type 'Route<Response<200, string, undefined>>' is not assignable to type 'Route<Response<200, string, { 'Content-Type': "application/json"; }>>'.