+++ typing-tests/tests/error-missing-response.ts
@@ -1,11 +1,12 @@
 import * as t from 'io-ts'
-import { Parser, Response, RouteHandler, routeHandler } from 'typera-koa'
+import { Parser, Response, Route, route } from 'typera-koa'
 
-export const handler: RouteHandler<Response.Ok<string>> = routeHandler(
-  Parser.body(t.type({ foo: t.number }))
-)(_req => {
-  return Response.ok('foo')
-})
+export const handler: Route<Response.Ok<string>> = route
+  .post('/')
+  .use(Parser.body(t.type({ foo: t.number })))
+  .handler(_req => {
+    return Response.ok('foo')
+  })
 
 // Expected error:
-// Type 'RouteHandler<RequestBase, Response<400, string, undefined> | Response<200, string, undefined>>' is not assignable to type 'RouteHandler<RequestBase, Response<200, string, undefined>>'.
+// Type 'Route<Response<400, string, undefined> | Response<200, string, undefined>>' is not assignable to type 'Route<Response<200, string, undefined>>'.