+++ typing-tests/tests/error-no-response.ts
@@ -1,6 +1,6 @@
-import { routeHandler } from 'typera-koa'
+import { route } from 'typera-koa'
 
-routeHandler()(_req => {})
+route.get('/').handler(_req => {})
 
 // Expected error:
 // Type 'void' is not assignable to type 'Response<number, any, OptionalHeaders> | Promise<Response<number, any, OptionalHeaders>>