+++ typing-tests/tests/error-wrong-response-body-type.ts
@@ -1,10 +1,10 @@
-import { Response, RouteHandler, routeHandler } from 'typera-koa'
+import { Response, Route, route } from 'typera-koa'
 
-export const handler: RouteHandler<Response.Ok<string>> = routeHandler()(
-  _req => {
+export const handler: Route<Response.Ok<string>> = route
+  .get('/')
+  .handler(_req => {
     return Response.ok(123)
-  }
-)
+  })
 
 // Expected error:
-// Type 'RouteHandler<RequestBase, Response<200, number, undefined>>' is not assignable to type 'RouteHandler<RequestBase, Response<200, string, undefined>>'.
+// Type 'Route<Response<200, number, undefined>>' is not assignable to type 'Route<Response<200, string, undefined>>'.