+++ typing-tests/tests/ok-async-route-handler.ts
@@ -1,7 +1,7 @@
 import * as typera from 'typera-koa'
 
-export const handler: typera.RouteHandler<typera.Response.Ok<
-  string
->> = typera.routeHandler()(async _req => {
-  return typera.Response.ok('foo')
-})
+export const root: typera.Route<typera.Response.Ok<string>> = typera.route
+  .get('/')
+  .handler(async _req => {
+    return typera.Response.ok('foo')
+  })