+++ typing-tests/tests/ok-parse-request.ts
@@ -1,25 +1,27 @@
-import { Response, Parser, RouteHandler, routeHandler } from 'typera-koa'
+import { Response, Parser, Route, URL, route } from 'typera-koa'
 import * as t from 'io-ts'
 
 interface ResponseBody {
-  x: string
+  x: number
   y: boolean
   z: string
   w: number
 }
 
-export const handler: RouteHandler<
+export const root: Route<
   Response.Ok<ResponseBody> | Response.NotFound | Response.BadRequest<string>
-> = routeHandler(
-  Parser.routeParams(t.type({ x: t.string })),
-  Parser.query(t.type({ y: t.boolean })),
-  Parser.headers(t.type({ z: t.string })),
-  Parser.body(t.type({ w: t.number }))
-)(req => {
-  return Response.ok({
-    x: req.routeParams.x,
-    y: req.query.y,
-    z: req.headers.z,
-    w: req.body.w,
+> = route
+  .get('/', URL.int('x'))
+  .use(
+    Parser.query(t.type({ y: t.boolean })),
+    Parser.headers(t.type({ z: t.string })),
+    Parser.body(t.type({ w: t.number }))
+  )
+  .handler(req => {
+    return Response.ok({
+      x: req.routeParams.x,
+      y: req.query.y,
+      z: req.headers.z,
+      w: req.body.w,
+    })
   })
-})