+++ typing-tests/tests/ok-response.ts
@@ -1,7 +1,7 @@
 import * as typera from 'typera-koa'
 
-export const handler: typera.RouteHandler<typera.Response.Ok<
-  number
->> = typera.routeHandler()(_req => {
+export const handler: typera.Route<
+  typera.Response.Ok<number>
+> = typera.route.get('/').handler(_req => {
   return typera.Response.ok(42)
 })